issue_7: RestaurantListクラスをコンポーネントファイルに分割
